Reorganize exportable assets classes.
